pyproject.toml
[project]
name = "cpg-poc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.7",
    "fastapi>=0.135.3",
    "jupyter>=1.1.1",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.13",
    "langgraph>=1.1.6",
    "pandas>=3.0.2",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.4.1",
]
